+++ src/main/java/team4/footwithme/stadium/service/response/StadiumDetailResponse.java
@@ -1,4 +1,4 @@
-package team4.footwithme.stadium.api.response;
+package team4.footwithme.stadium.service.response;
 
 public record StadiumDetailResponse(
         Long memberId,
