+++ src/main/java/team4/footwithme/stadium/service/response/StadiumsResponse.java
@@ -1,3 +1,3 @@
-package team4.footwithme.stadium.api.response;
+package team4.footwithme.stadium.service.response;
 
 public record StadiumsResponse(Long stadiumId, String name, String address) {}
